modified the custom hook that handles breakpoints
