+++ src/components/PageHeader/StatsBar/index.tsx
@@ -5,7 +5,7 @@ const StatsBar = async () => {
   const data = await getStatsBarData();
 
   return (
-    <div className="flex h-14 w-full justify-center gap-8 bg-[var(--secondary-foreground)] text-xs text-[#ffffff;]">
+    <div className="flex h-14 w-full justify-center gap-2 bg-[var(--secondary-foreground)] text-xs text-[#ffffff;] md:gap-8">
       <StatsBarList data={data} />
     </div>
   );
